third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/security/contentSecurityPolicy/1.1/scripthash-handler-blocked-expected.txt

CONSOLE ERROR: line 10: Refused to execute inline event handler because it violates the following Content Security Policy directive: "script-src 'sha256-nhtYaXCssBJTThiDLYewspQYue9tisulEwJ3nTJKcMI='". Either the 'unsafe-inline' keyword, a hash ('sha256-...'), or a nonce ('nonce-...') is required to enable inline execution.

This is a testharness.js-based test.
PASS Inline event handlers whitelisted by the policy should not fire, as 'unsafe-hash-attributes' is not present. 
Harness: the test ran to completion.
